// checkout/src/legacy-api/enum/payment-method.enum.ts
export enum PaymentMethodEnum {
  METHOD_SANTANDER_DE_INSTALLMENT = 'santander_installment',
  METHOD_SANTANDER_DE_INVOICE = 'santander_invoice_de',
  METHOD_SANTANDER_DE_POS_INVOICE = 'santander_pos_invoice_de',
  METHOD_SANTANDER_DE_FACTORING = 'santander_factoring_de',
  METHOD_SANTANDER_DE_POS_FACTORING = 'santander_pos_factoring_de',
  METHOD_SANTANDER_DE_CCP_INSTALLMENT = 'santander_ccp_installment',
  METHOD_SANTANDER_DE_POS_INSTALLMENT = 'santander_pos_installment',
  METHOD_SANTANDER_NO_INSTALLMENT = 'santander_installment_no',
  METHOD_SANTANDER_NO_POS_INSTALLMENT = 'santander_pos_installment_no',
  METHOD_SANTANDER_NO_INVOICE = 'santander_invoice_no',
  METHOD_SANTANDER_NO_POS_INVOICE = 'santander_pos_invoice_no',
  METHOD_SANTANDER_DK_INSTALLMENT = 'santander_installment_dk',
  METHOD_SANTANDER_DK_POS_INSTALLMENT = 'santander_pos_installment_dk',
  METHOD_SANTANDER_SE_INSTALLMENT = 'santander_installment_se',
  METHOD_SANTANDER_SE_POS_INSTALLMENT = 'santander_pos_installment_se',
  METHOD_SANTANDER_AT_INSTALLMENT = 'santander_installment_at',
  METHOD_SANTANDER_NL_INSTALLMENT = 'santander_installment_nl',
  METHOD_SANTANDER_UK_INSTALLMENT = 'santander_installment_uk',
  METHOD_SANTANDER_UK_POS_INSTALLMENT = 'santander_pos_installment_uk',
  METHOD_SANTANDER_FI_INSTALLMENT = 'santander_installment_fi',
  METHOD_SANTANDER_FI_POS_INSTALLMENT = 'santander_pos_installment_fi',
  METHOD_SANTANDER_BE_INSTALLMENT = 'santander_installment_be',

  METHOD_INVOICE = 'invoice',
  METHOD_CASH = 'cash',
  METHOD_SOFORT = 'sofort',
  METHOD_PAYMILL_CREDIT_CARD = 'paymill_creditcard',
  METHOD_PAYMILL_DIRECT_DEBIT = 'paymill_directdebit',
  METHOD_PAYPAL = 'paypal',
  METHOD_STRIPE_CREDIT_CARD = 'stripe',
  METHOD_STRIPE_DIRECT_DEBIT = 'stripe_directdebit',
  METHOD_PAYEX_FAKTURA = 'payex_faktura',
  METHOD_PAYEX_CREDIT_CARD = 'payex_creditcard',
  METHOD_APPLE_PAY = 'apple_pay',
  METHOD_GOOGLE_PAY = 'google_pay',
  METHOD_SWEDBANK_INVOICE = 'swedbank_invoice',
  METHOD_SWEDBANK_CREDITCARD = 'swedbank_creditcard',
  METHOD_ZINIA_BNPL = 'zinia_bnpl',
  METHOD_ZINIA_POS = 'zinia_pos',
  METHOD_ZINIA_INSTALLMENT = 'zinia_installment',
  METHOD_ZINIA_POS_INSTALLMENT = 'zinia_pos_installment',
  METHOD_ZINIA_SLICE_THREE = 'zinia_slice_three',
  METHOD_ZINIA_POS_SLICE_THREE = 'zinia_pos_slice_three',
  METHOD_ZINIA_BNPL_DE = 'zinia_bnpl_de',
  METHOD_ZINIA_POS_DE = 'zinia_pos_de',
  METHOD_ZINIA_INSTALLMENT_DE = 'zinia_installment_de',
  METHOD_ZINIA_POS_INSTALLMENT_DE = 'zinia_pos_installment_de',
  METHOD_ZINIA_SLICE_THREE_DE = 'zinia_slice_three_de',
  METHOD_ZINIA_POS_SLICE_THREE_DE = 'zinia_pos_slice_three_de',
  METHOD_ZINIA_LENDING_DE = 'zinia_lending_de',
  METHOD_ZINIA_POS_LENDING_DE = 'zinia_pos_lending_de',
  METHOD_IVY = 'ivy',
  METHOD_IDEAL = 'ideal',
  METHOD_VIPPS = 'vipps',
  METHOD_ALLIANZ = 'allianz_trade_b2b_bnpl',
  METHOD_SWISH = 'swish',
  METHOD_MOBILE_PAY = 'mobile_pay',
  METHOD_TRUSTLY = 'trustly',
  METHOD_SANTANDER_B2B = 'psa_b2b_bnpl',
  METHOD_INSTANT_PAYMENT = 'instant_payment',
  METHOD_NETS = 'nets',
  METHOD_BFS = 'bfs_b2b_bnpl',
  METHOD_HSBC = 'hsbc',
  METHOD_SANTANDER_INSTANT_AT = 'santander_instant_at',
}